requirements.txt
gradio
assemblyai
together
Pillow
PyYAML